+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Surround auto-saved SSH session closed markers with blank lines to improve visibility.
 - Increase log panel SSH message height to improve visibility.
 - Update log message handling behavior. Messages from auto-save or from logger commands are handled in a secondary line from the connection status.
+- When loglevel filter is set to ALL, do not hide blank lines and lines without a log tag.
 
 ### Deprecated
 - None.